+++ DesignToolbox/DesignToolbox/Pages/Components/Navigations/ToolBar/ToolBarTop/ToolBarTopConfigurationView.swift
@@ -27,10 +27,6 @@ final class ToolBarTopConfigurationModel: ToolBarConfigurationModel {
         didSet { updateCode() }
     }
 
-    @Published var title: String {
-        didSet { updateCode() }
-    }
-
     @Published var largeTitle: Bool {
         didSet { updateCode() }
     }
@@ -52,7 +48,6 @@ final class ToolBarTopConfigurationModel: ToolBarConfigurationModel {
     override init() {
         demoOption = .navigation
 
-        title = "app_components_topAppBar_title_label".localized()
         largeTitle = false
         subTitle = ""
 
@@ -92,7 +87,7 @@ final class ToolBarTopConfigurationModel: ToolBarConfigurationModel {
     // MARK: Code update
 
     private var titlePattern: String {
-        !title.isEmpty ? "\"\(title)\"," : ""
+        !title.isEmpty ? "\"\(title)\", " : ""
     }
 
     private var hasLargeTitlePattern: String {
@@ -110,6 +105,10 @@ final class ToolBarTopConfigurationModel: ToolBarConfigurationModel {
               \(leadingItemsPattern)
             }
             """
+        let principal = principalType != .none ?
+            """
+            , principalItem: \(principalItemPattern)
+            """ : ""
         let trailing = trailingItemPattern.isEmpty ? "" :
             """
             , trailingItems: {
@@ -119,7 +118,7 @@ final class ToolBarTopConfigurationModel: ToolBarConfigurationModel {
 
         code = """
         SomeView()
-        .toolBarTop(\(titlePattern)\(hasLargeTitlePattern)\(subtitlePattern)\(leading)\(trailing))
+        .toolBarTop(\(titlePattern)\(hasLargeTitlePattern)\(subtitlePattern)\(leading)\(principal)\(trailing))
         """
     }
 }
@@ -149,16 +148,17 @@ struct ToolBarTopConfiguration: View {
                     OUDSSwitchItem("app_components_topAppBar_hideBackButton_tech", isOn: $configurationModel.hideBackButton)
                 }
 
+                OUDSHorizontalDivider()
+
                 ToolBarLeadingConfiguration(configurationModel: configurationModel)
 
-                ToolBarTrailingConfiguration(configurationModel: configurationModel)
-                if configurationModel.trailing == .icon {
-                    OUDSHorizontalDivider()
+                OUDSHorizontalDivider()
 
-                    OUDSChipPicker(title: "app_components_badge_tech".localized(),
-                                   selection: $configurationModel.badgeType,
-                                   chips: BarItemBadgeType.chips)
-                }
+                ToolBarPrincipalConfiguration(configurationModel: configurationModel)
+
+                OUDSHorizontalDivider()
+
+                ToolBarTrailingConfiguration(configurationModel: configurationModel)
 
                 ToolBarItemStyle(configurationModel: configurationModel)